Refactor Dataframe (for WebWorkers)
